requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
redis
httpx
asyncpg
pydantic
pymongo
python-dotenv

pytest
pytest-mock
pytest-cov
faker
freezegun

apache-airflow-providers-postgres
apache-airflow-providers-mongo
apache-airflow-providers-http
apache-airflow-providers-redis
apache-airflow-providers-amazon
apache-airflow-providers-google
apache-airflow-providers-slack

dbt-core
dbt-bigquery
dbt-postgres
# protobuf
# opentelemetry-proto
astronomer-cosmos==1.10.2